requirements.txt
flask==2.2.5
werkzeug==2.2.3
mysql-connector-python==8.0.26
pandas==1.3.3
numpy==1.23.5
tensorflow==2.12.0
tensorflow-recommenders>=0.7.0
scikit-learn==1.2.2
azure-ai-inference==1.0.0b4
tabulate==0.9.0
gunicorn==20.1.0
